protect resources with authorization

+++ backend/routes/posts.js
@@ -60,8 +60,13 @@ router.put("/:id",checkAuth, multer({storage: storage}).single("image"), (req, r
         imagePath: imagePath
     });
     console.log(post);
-    Post.updateOne({_id: req.params.id}, post).then(result => {
-        res.status(200).json({ message: "Update successful!" });
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+        console.log(result);
+        if(result.nModified > 0){
+            res.status(200).json({ message: "Update successful!" });
+        }else{
+            res.status(401).json({ message: "Not authorized!" });
+        }
     });
 });
 
@@ -97,9 +102,13 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id",checkAuth, (req, res, next) => {
-    Post.deleteOne({_id: req.params.id}).then((result) => {
+    Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then((result) => {
         console.log(result);
-        res.status(200).json({message: "Post deleted!"});
+        if(result.n > 0){
+            res.status(200).json({ message: "Deletion successful!" });
+        }else{
+            res.status(401).json({ message: "Not authorized!" });
+        }
     });
 });
